+++ packages/router/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @tinyhttp/router
 
+## 0.0.2
+
+### Patch Changes
+
+- Fix minor issues in the router
+
 ## 0.0.1
+
 ### Patch Changes
 
 - Split app into req, res, send and router